# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the CIJ project",
        "title": "CIJ Project API",
        "contact": {
            "name": "API Support",
            "email": "cauakathdev@gmail.com"
        },
        "version": "1.0"
    },
    "host": "conexao-inclusao.com",
    "basePath": "/",
    "paths": {
...
        "/config/{email}": {
            "get": {
                "description": "Get user config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Get user config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
        "/news": {
            "get": {
                "description": "list all registered news.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "List all registered news.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.NewsResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new news.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "Create a new news.",
                "parameters": [
                    {
                        "type": "string",
                        "name": "author",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "date",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "banner",
                        "name": "banner",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "author_image",
                        "name": "authorImage",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "enum.ColorBlindnessEnum": {
            "type": "string",
            "enum": [
                "normal",
                "protanopia",
                "deuteranopia",
                "tritanopia"
            ],
            "x-enum-varnames": [
                "Normal",
                "Protanopia",
                "Deuteranopia",
                "Tritanopia"
            ]
        },
...
        "enum.ThemeEnum": {
            "type": "string",
            "enum": [
                "light",
                "dark",
                "system"
            ],
            "x-enum-varnames": [
                "Light",
                "Dark",
                "System"
            ]
        },
...
        "model.Config": {
            "type": "object",
            "properties": {
                "color_blindness": {
                    "$ref": "#/definitions/enum.ColorBlindnessEnum"
                },
                "font_size": {
                    "type": "integer"
                },
                "screen_reader": {
                    "type": "boolean"
                },
                "system_colors": {
                    "$ref": "#/definitions/model.SystemColors"
                },
                "theme": {
                    "$ref": "#/definitions/enum.ThemeEnum"
                },
                "voice_capture": {
                    "type": "boolean"
                }
            }
        },
...
        "model.Field": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "model.NewsResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "author_image": {
                    "type": "string"
                },
                "banner": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Field"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.SystemChartColors": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "model.SystemColors": {
            "type": "object",
            "properties": {
                "chart_colors": {
                    "$ref": "#/definitions/model.SystemChartColors"
                },
                "primary_colors": {
                    "$ref": "#/definitions/model.SystemPrimaryColors"
                }
            }
        },
        "model.SystemPrimaryColors": {
            "type": "object",
            "properties": {
                "background_color": {
                    "type": "string"
                },
                "font_color": {
                    "type": "string"
                },
                "input_color": {
                    "type": "string"
                },
                "primary_color": {
                    "type": "string"
                },
                "secondary_color": {
                    "type": "string"
                },
                "secondary_font_color": {
                    "type": "string"
                }
            }
        },
...
    }
}